fixed a "comparison always FALSE/TRUE due to limited ..."
git-svn-id: https://svn.aros.org/svn/aros/trunk@8530 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/aminet/comm/term/MicroTerm/console.c
@@ -170,7 +170,7 @@ struct MySer *ms;
         if (msg->Class == MENUPICK)
         {
           num = msg->Code;
-          while (num != MENUNULL)
+          while (num != (short)MENUNULL)
           {
             switch (MENUNUM(num))
             {
